+++ src/database/neighbor_table/nt.h
@@ -36,7 +36,7 @@ int db_nt_init();
 
 int db_nt_reset_rssi(uint8_t ether_neighbor_addr[ETH_ALEN], dessert_meshif_t* iface, struct timeval* timestamp);
 
-int db_nt_update_rssi(uint8_t ether_neighbor[ETH_ALEN], dessert_meshif_t* iface, struct timeval* timestamp);
+int8_t db_nt_update_rssi(uint8_t ether_neighbor[ETH_ALEN], dessert_meshif_t* iface, struct timeval* timestamp);
 /**
  * Take a record that the given neighbor seems to be
  * 1 hop bidirectional neighbor
